% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiciandata-package.R
\docType{package}
\name{physiciandata-package}
\alias{physiciandata}
\alias{physiciandata-package}
\title{physiciandata: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrewallenbruce.github.io/physiciandata/}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewbruce.himni@gmail.com} (\href{https://orcid.org/0000-0001-5040-4327}{ORCID})

}
\keyword{internal}
